[chore] DBITests 파일 옮기기
